pyproject.toml
[tool.poetry]
name = "cookplanner"
version = "0.1.0"
description = "Mattat kindergarten cooking scheduler"
authors = ["Shahar Evron"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
PyYAML = "^5.4.1"
google-api-python-client = "^2.19.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mypy = "^0.910"
flake8 = "^3.9.2"
black = "^21.8b0"
isort = "^5.9.3"
pytest-icdiff = "^0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
